Rename danger.yml to danger.yml
Your GitHub Actions workflow is well-structured, with some aspects that need to be improved. There are some potential issues and improvements which can be discussed below, such as; actions / checkout Version Pinning It is recommended to pin the actions/checkout action at a given version number(eg:v2), instead of using@master, in order for it be stable. actions / setup-node Version Pinning For sanity reasons, it is advisable to always pin the actions/setup-node action at a given version eg.v3 rather than using latest".